+++ src/libs/BoardManagement.js
@@ -18,7 +18,7 @@ class BoardManagement {
 
   changeVisibilityBoard(boardId, newVisibility){
     // console.log(this.boards[this.boards.findIndex(board => board.id === boardId)].visibility)
-    this.boards[this.boards.findIndex(board => board.id === boardId)].visibility = newVisibility
+    this.boards.personalBoards[this.boards.personalBoards.findIndex(board => board.id === boardId)].visibility = newVisibility
     // console.log(this.boards[this.boards.findIndex(board => board.id === boardId)].visibility)
   }
